"8:45pm. Nothing. Let me me use it just as a bias.
...It is halfway decent, but worse than using it as a hyperparameter.

Let me go back to using the `abs n` update.

It works a bit better, but still worse than the hyperparam one. Let me set n to a singleton.

8:55am.

```
The learning rate is 2 ** -14.5
At iteration 9 the cost is 26.1772171020508
At iteration 19 the cost is 25.0954975128174
At iteration 29 the cost is 23.6455934524536
At iteration 39 the cost is 22.8536647796631
At iteration 49 the cost is 20.5001537322998
At iteration 59 the cost is 18.1396229743958
At iteration 69 the cost is 16.6236804962158
At iteration 79 the cost is 19.1437189102173
At iteration 89 the cost is 14.3266639709473
At iteration 99 the cost is 11.2737712144852
At iteration 109 the cost is 13.0005639314651
At iteration 119 the cost is 10.7498154640198
At iteration 129 the cost is 11.0078866899014
At iteration 139 the cost is 11.6846738815308
At iteration 149 the cost is 7.8558707177639
At iteration 159 the cost is 10.8808064579964
At iteration 169 the cost is 10.848389840126
At iteration 179 the cost is 10.6550312995911
At iteration 189 the cost is 7.28957840204239
At iteration 199 the cost is 6.34652535915375
At iteration 209 the cost is 9.05903337001801
At iteration 219 the cost is 6.07122496366501
At iteration 229 the cost is 8.01146991252899
At iteration 239 the cost is 4.58705169856548
At iteration 249 the cost is 8.39310302734375
At iteration 259 the cost is 4.63726053237915
At iteration 269 the cost is 5.82010776996613
At iteration 279 the cost is 7.93234586417675
At iteration 289 the cost is 6.61548615694046
At iteration 299 the cost is 5.32437694072723
At iteration 309 the cost is 8.71889964044094
```

Backprop just is not good at optimizing this sort of thing. It just causes instability.

Let me do one more test with the `abs n` version and then I am done with this.

...Actually the thing I was testing was the `abs n` version.

9am. Crap. It is all crap.

Forget it. This is a failure. It is really hard to predict how various components will work in advance it seems. I did not expect anything from plasticity and it turned out to be huge. I expected this to be a gain and it turned out to be a flop.

Let me go to the next thing on the list. I need to investigate the outputs. I need to find out if the way I am passing them into the update right now is right or if I would benefit more by separating them and allowing the net a wider range of learning. Assuming that this works, I might want to try using the MI in the head of the layer again."