"9:10am.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 25.9692358016968
At iteration 19 the cost is 23.8122661590576
At iteration 29 the cost is 20.888045501709
At iteration 39 the cost is 20.3738376617432
At iteration 49 the cost is 16.4379911422729
At iteration 59 the cost is 13.5112490653992
At iteration 69 the cost is 13.1530016899109
At iteration 79 the cost is 11.8732012748718
At iteration 89 the cost is 9.3266978263855
At iteration 99 the cost is 7.47329640388489
At iteration 109 the cost is 10.0621089935303
At iteration 119 the cost is 4.80078272819519
At iteration 129 the cost is 7.17729148864746
At iteration 139 the cost is 7.04925043582916
At iteration 149 the cost is 4.19620559215546
At iteration 159 the cost is 6.45375134944916
At iteration 169 the cost is 7.6301051735878
At iteration 179 the cost is 6.16887801885605
At iteration 189 the cost is 6.42746723294258
At iteration 199 the cost is 5.3134494304657
At iteration 209 the cost is 5.88463004827499
At iteration 219 the cost is 4.45720526725054
At iteration 229 the cost is 6.37673371434212
At iteration 239 the cost is 7.80014255046845
At iteration 249 the cost is 7.56163257956505
At iteration 259 the cost is 6.17873280346394
At iteration 269 the cost is 6.85374718904495
At iteration 279 the cost is 9.36737926006317
At iteration 289 the cost is 6.83577544689178
At iteration 299 the cost is 5.20602827072144
At iteration 309 the cost is 9.97592759728432
```

Let me lower the learning rate.

```
The learning rate is 2 ** -13.5
At iteration 9 the cost is 26.3370258331299
At iteration 19 the cost is 25.2293458938599
At iteration 29 the cost is 23.8267389297485
At iteration 39 the cost is 23.6890790939331
At iteration 49 the cost is 22.0681831359863
At iteration 59 the cost is 20.5225505828857
At iteration 69 the cost is 20.1777429580688
At iteration 79 the cost is 19.6079933166504
At iteration 89 the cost is 17.4009021759033
At iteration 99 the cost is 15.7142977714539
At iteration 109 the cost is 15.952345085144
At iteration 119 the cost is 12.9569246292114
At iteration 129 the cost is 12.1752437591553
At iteration 139 the cost is 11.316818189621
At iteration 149 the cost is 9.04496140480041
At iteration 159 the cost is 10.0486815452576
At iteration 169 the cost is 8.95414319038391
At iteration 179 the cost is 8.18642220497131
At iteration 189 the cost is 7.38001482486725
At iteration 199 the cost is 7.0346312046051
At iteration 209 the cost is 6.86756081581116
At iteration 219 the cost is 5.23268712759018
At iteration 229 the cost is 5.21665465831757
At iteration 239 the cost is 5.51858979463577
At iteration 249 the cost is 6.18612926006317
At iteration 259 the cost is 4.84413438439369
At iteration 269 the cost is 6.68087933063507
At iteration 279 the cost is 8.45523886680603
At iteration 289 the cost is 5.88826471567154
At iteration 299 the cost is 4.63561503887177
At iteration 309 the cost is 9.79597836732864
At iteration 319 the cost is 4.89717614650726
At iteration 329 the cost is 5.7616689324379
At iteration 339 the cost is 3.96988346576691
At iteration 349 the cost is 5.95464142560959
At iteration 359 the cost is 6.87406200170517
At iteration 369 the cost is 6.3441597700119
At iteration 379 the cost is 6.80766094923019
At iteration 389 the cost is 4.7083674788475
At iteration 399 the cost is 5.05200781226158
At iteration 409 the cost is 3.68862997293472
At iteration 419 the cost is 6.42802909612656
At iteration 429 the cost is 13.4660689353943
At iteration 439 the cost is 5.06943873763084
At iteration 449 the cost is 6.13699641227722
At iteration 459 the cost is 5.44074866771698
At iteration 469 the cost is 5.99045652151108
At iteration 479 the cost is 4.42189815044403
At iteration 489 the cost is 5.18628816604614
At iteration 499 the cost is 5.26931860744953
At iteration 509 the cost is 5.16473652720451
At iteration 519 the cost is 3.08573217391968
```

No, this is not good enough.

9:15am. Another failure. That just leave the plastic LSTM. But let first check out if varying n does anything significant."