"10:30am. Right. Let me just do it. I was thinking about how to do a multiscale plastic LSTM, but I think that would be overkill. I guess it would be worth a try. Trying out various architecture will be my mission for the day. Tomorrow I will implement optimized implementations and after that I will look into combining what I have here with PRONG.
I'll make a new implementation of PRONG that uses buffered updates suitable for use in RL. And as annoying as those biases are, I am going to have to give them the same treatmnet as for everything else."

I got distracted for a bit. Let me get back into the zone of things.