"1:55pm. It does not work. This is giving me some pause. I am thinking right now.
It occurs to me that there is a far easier way of implementing what I want and that would be to make n a range rather than a rigid value.

That way this could work is that I could assign each neuron a static learning rate from slowest to fastest. That is a really good idea, but before that, let me finish with what I had in mind now. This should also work and the fact that it does not is reason enough for me to switch gears. I'll go with what was my actual idea."