"2:55pm.
```
                    inl ins =
                        inl a = Struct.choose id (adjoints ins)
                        inl b = index_into dim a
                        Struct.map2 (inl a b -> b) a b // This is to get rid of the `n` in a safe way.

                    Struct.iter3 (inl in b f -> f in (error * b ())) ins ads output_functions
```

Getting rid of n was a bit tricky, but I got this to compile. Let me give it a try.

2:55pm.

```
            multiscale_v1 =
                Modulated.multiscale_v1 ({from=0.003f32; near_to=0.03f32}) size.pattern
            }
        }

    init s size.pattern network.modulated.multiscale_v1
```

It does not work completely well just yet. Let me try in multiscale. This is just one n. Since it rescales the output before the tanh, it might do better here.

3pm.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 27.138786315918
At iteration 19 the cost is 23.8437343597412
At iteration 29 the cost is 20.8172521591187
At iteration 39 the cost is 19.8896774291992
At iteration 49 the cost is 17.3276117324829
At iteration 59 the cost is 15.1898682594299
At iteration 69 the cost is 14.9248244285583
At iteration 79 the cost is 15.2801350593567
At iteration 89 the cost is 13.0958507537842
At iteration 99 the cost is 11.0106127738953
At iteration 109 the cost is 12.2647009849548
At iteration 119 the cost is 9.70887212753296
At iteration 129 the cost is 9.87632970809937
At iteration 139 the cost is 9.57316427230835
At iteration 149 the cost is 8.43969917297363
At iteration 159 the cost is 9.2598735332489
At iteration 169 the cost is 8.98571310043335
At iteration 179 the cost is 8.68534836769104
At iteration 189 the cost is 7.99626183509827
At iteration 199 the cost is 7.73200254440308
At iteration 209 the cost is 7.54712243080139
At iteration 219 the cost is 6.73473815917969
At iteration 229 the cost is 6.96710867881775
At iteration 239 the cost is 6.26500234603882
At iteration 249 the cost is 7.12966451644897
At iteration 259 the cost is 6.13381853103638
At iteration 269 the cost is 6.92004051208496
At iteration 279 the cost is 7.01241068840027
At iteration 289 the cost is 6.39029159545898
At iteration 299 the cost is 5.83351187705994
At iteration 309 the cost is 6.87587447166443
At iteration 319 the cost is 5.909077501297
At iteration 329 the cost is 5.68367228507996
At iteration 339 the cost is 5.32983169555664
At iteration 349 the cost is 5.73273270130157
At iteration 359 the cost is 6.31375637054443
At iteration 369 the cost is 5.78543033599854
At iteration 379 the cost is 5.96601815223694
```

Hmmm, was I wrong?

3pm.

```
            multiscale_v1 =
                Modulated.multiscale_v1 ({from=n; near_to=n; block=()}) size.pattern
```

No, it works worse even with this. I might be encountering some kind of bug.

3:05pm.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 25.7930891036987
At iteration 19 the cost is 23.2897783279419
At iteration 29 the cost is 20.2744428634644
At iteration 39 the cost is 19.74271068573
At iteration 49 the cost is 16.3448669433594
At iteration 59 the cost is 13.5301927566528
At iteration 69 the cost is 13.4036648750305
At iteration 79 the cost is 12.6869342803955
At iteration 89 the cost is 9.77825689315796
At iteration 99 the cost is 7.83521146774292
At iteration 109 the cost is 9.03099756240845
```

This is the vanilla oja with a scalar n.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 25.7930891036987
At iteration 19 the cost is 23.2897783279419
At iteration 29 the cost is 20.2744428634644
At iteration 39 the cost is 19.7427104949951
At iteration 49 the cost is 16.3448666572571
At iteration 59 the cost is 13.530192899704
At iteration 69 the cost is 13.4036652565002
At iteration 79 the cost is 12.6869342803955
At iteration 89 the cost is 9.77825689315796
At iteration 99 the cost is 7.8352114200592
At iteration 109 the cost is 9.03099765777588
```

This is with a range n. So they are roughly the same. Ok.

3:05pm. So was that multiscale thing my imagination?"