"3:20pm. Things are not improving by implementing ranged learning rates, but this is just because the particular problem I am facing is special. The reason why it is making things unstable is because the outputs have to be precise in the output. On a standard task this would not be the case, but here it is. So it is important the the outputs at a partular location have the necessary capacity.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 22.4548286437988
At iteration 19 the cost is 15.875200176239
At iteration 29 the cost is 11.4388102293015
At iteration 39 the cost is 9.46172976493835
At iteration 49 the cost is 6.19859046936035
At iteration 59 the cost is 3.78028498888016
At iteration 69 the cost is 3.78395716547966
At iteration 79 the cost is 5.28113031387329
At iteration 89 the cost is 3.72839379310608
At iteration 99 the cost is 2.3129522331059
At iteration 109 the cost is 3.24694517254829
At iteration 119 the cost is 1.75639936625957
At iteration 129 the cost is 1.71839220225811
At iteration 139 the cost is 2.19462642967701
At iteration 149 the cost is 0.806012111902237
At iteration 159 the cost is 2.69796884804964
At iteration 169 the cost is 2.20530000030994
At iteration 179 the cost is 2.26737449169159
At iteration 189 the cost is 1.42291231155396
At iteration 199 the cost is 1.28987006880343
```

Ok, enough of this, I do not feel like waiting anymore as it is so slow. This is with 3 cells all with the `n=0.01` learning rate. It is fine.

As I said, in a real task I would have a lot more freedom. I could simply use a wider layer than this weird attention thing.

3:25pm. The next comes the LSTM. This multicell plastic RNN is actually decently close to it. I just need to add an extra layer and a memory cell. That will take care of it."