"3:45pm.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 23.5378101348877
At iteration 19 the cost is 18.6510457038879
At iteration 29 the cost is 13.8500819206238
At iteration 39 the cost is 12.4917402267456
At iteration 49 the cost is 8.60422971248627
At iteration 59 the cost is 5.83245108127594
At iteration 69 the cost is 5.69557902812958
At iteration 79 the cost is 6.50192060470581
At iteration 89 the cost is 4.45403064489365
At iteration 99 the cost is 2.78473097980022
At iteration 109 the cost is 4.09673142433167
At iteration 119 the cost is 1.97785163521767
At iteration 129 the cost is 2.44129035174847
At iteration 139 the cost is 2.56764996647835
At iteration 149 the cost is 1.27775567471981
At iteration 159 the cost is 2.71358213424683
At iteration 169 the cost is 2.58357965350151
At iteration 179 the cost is 2.38486142084002
At iteration 189 the cost is 1.55882948487997
At iteration 199 the cost is 1.41664651781321
At iteration 209 the cost is 1.03303005695343
At iteration 219 the cost is 0.795522707700729
At iteration 229 the cost is 1.26028157174587
At iteration 239 the cost is 0.887175200879574
At iteration 249 the cost is 2.80705315470696
At iteration 259 the cost is 1.03316257763654
At iteration 269 the cost is 2.16772762089968
At iteration 279 the cost is 1.95873377472162
At iteration 289 the cost is 2.23742924612016
At iteration 299 the cost is 0.689861108362675
At iteration 309 the cost is 4.8141966894269
At iteration 319 the cost is 1.09730467423797
At iteration 329 the cost is 0.644836599379778
At iteration 339 the cost is 0.604679702967405
At iteration 349 the cost is 1.33287801779807
At iteration 359 the cost is 1.94042064845562
At iteration 369 the cost is 0.833737899363041
At iteration 379 the cost is 2.00245912224054
At iteration 389 the cost is 0.83235881254077
At iteration 399 the cost is 0.667068579792976
At iteration 409 the cost is 1.17754584401846
At iteration 419 the cost is 1.99419609755278
At iteration 429 the cost is 2.4874530851841
At iteration 439 the cost is 2.43623539954424
At iteration 449 the cost is 1.008605286479
At iteration 459 the cost is 1.76994005739689
At iteration 469 the cost is 2.9883710578084
At iteration 479 the cost is 0.605719475448132
At iteration 489 the cost is 2.35091164559126
At iteration 499 the cost is 2.90004221014678
At iteration 509 the cost is 2.01719920188189
At iteration 519 the cost is 0.392204314470291
At iteration 529 the cost is 1.11258580610156
At iteration 539 the cost is 0.413752922415733
```

I do not feel like letting run all the way, but removing that tanh in the middle layers did improve things. Wonderful.

3:45pm. Let me take a short break here. When I get back I'll start work on the plastic LSTM."