"7:35am. I am up early today.
7:40am. Let me start programming.

https://www.reddit.com/r/reinforcementlearning/comments/9gc84e/nvidia_ntech_september_2018_talk_by_ilya/?st=jm7abttx&sh=08ef42e0

I'll take a look at the talk by Ilya later. Yesterday I was so exhausted that I just went straight to bed after watching Baki. These 3.5 years of constant marathon programming sessions are not a joke.

And now I am finally at the point when I am doing RL seriously. After I finish the LSTM, make an optimized implementation for it and merge it with PRONG I will have something particularly powerful in my hands - though it may turn out the case that I will just opt for vanilla RNNs in the end.

Optimized PRONGed plastic RNN is something I will be using in the end. I never expected that modulating the plasticity of the network would give such massive benefits.

It is true that I cannot get the squared error to zero, but thinking more on it maybe that is not the best measure to use. It might be better use the same measure as in the Uber test by Miconi which is the sign accuracy.

Because the learning here is semi-supervised the network might never have enough capacity to get it directly down to zero. When I am done with the LSTM, this is something I will check, but for now I am curious about how well the LSTM will do.

7:50am. So my goal for the next few hours will be completing the LSTM activation in a style that befits me.

I will finally bring out AD in an actual kernel.

Once I have an optimized PRONGed plastic RNN on my hands, I will finally have something that could be called a real accomplishment. I'll head into poker, but I will take a look at Kaggle to see if there are any interesting metalearning contests. If there are I will stand a good chance of winning with the new architecture. Winning 100k or even 10k dollars would be a great benefit to me."