"11am.
```
        inl lstm_activation memory_old {ins with input output forget memory} s =
            inl b,a as dim = memory_old.dim
            assert_dim = assert_dim (primals ins)
            Liple.iter (inl k -> assert_dim k dim) (.input,.output,.forget,.memory)
            inl in = {ins with memory_old}

            inl out =
                inl x = to_dev_tensor {in=primals in}
                s.CudaFun.map {dim} (inl dim -> CudaAD.lstm_activation x .out)
                |> Struct.map (dr s)

            {
            out
            bck=met _ ->
                inl x = to_dev_tensor {in out}
                s.CudaKernel.iter {dim} (inl dim -> CudaAD.lstm_activation x |> CudaAD.run)
            }
```

Wow, just wow. This is beyond good. The LSTM activation is a very difficult thing to do cleanly in its optimized form and yet here I've completely mastered it.

The first time I did an LSTM was in 2015 and it completely crushed me. I actually could not get it to work back when I was doing the backward pass by hand and it drove me to study more and do research on AD. When I discovered that function could be used as closures to store the backward pass that was one of my major crystallizing moments. It is the one that set me on the path of functional programming. Up to then I was programming the same way I did in high school.

And now I am here and I am doing this.

It is not over yet. I can still get better. There are levels that exist beyond the one I am at right now. I can still get better.

11:20am. I got distracted for a bit here. Let me see if I can get this to compile and then I will take a break...

No actually, let me stop here as I need to do some chores. I'll do this properly. Programming under time pressure never works well."