"5:55pm.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 70.0651420593262
At iteration 19 the cost is 45.2638778686523
At iteration 29 the cost is 41.4161636352539
At iteration 39 the cost is 43.984903717041
```

Nope. Well, this much is to be expected. Rarely anything works the first time. The LSTM might even be correct, but it might turn out to be the case that those gates are getting in the way on this particular problem. Also the multiscale RNN is not compiling, but I do not feel like dealing...actually, let me deal with that as the last thing today.

```
type (layout_heap ([{block=[]; cell={bias={bias=[lit 1i64, lit 50i64]; block=[]}; factor={block=[]; constant={dim=[lit 1i64, lit 50i64]; init=lit 1.000000f32}}; input={block=[]; dr={block=[]; identity=[lit 50i64, lit 50i64]}}; modulator={input={bias={c={bia
```

Probably the initializer is messed up.

6pm. Yeah, it is not wonder as I'd removed cell. Let me replace it with unwrap instead.

Rather than unwrap it should be have been wrap.

6:05pm.

```
                    inl cell init = {
                        wrap = {
                            input = Initializer.dr (Initializer.identity (sublayer_size, size))
                            state = Initializer.randn {stddev=0.01f32; dim=size, size}
                            modulator =
                                {
                                input = modulator()
                                state = modulator()
                                }
                            bias = Initializer.constant {dim=1,size; init}
                            }
                        block=()
                        }
```

Syntax error...

```
                    inl cell _ = {
                        wrap = {
                            {
                            input = Initializer.dr (Initializer.identity (sublayer_size, size))
                            state = Initializer.randn {stddev=0.01f32; dim=size, size}
                            modulator =
                                {
                                input = modulator()
                                state = modulator()
                                }
                            bias = Initializer.bias (1,size)
                            factor = Initializer.constant {dim=1,size; init=to float 1}
                            }
                        block=()
                        }
```

Ah, this is missing a closing paren.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 23.5378099441528
At iteration 19 the cost is 18.6510448455811
At iteration 29 the cost is 13.8500822067261
At iteration 39 the cost is 12.4917403697968
At iteration 49 the cost is 8.60422956943512
At iteration 59 the cost is 5.83245100975037
At iteration 69 the cost is 5.69557890892029
At iteration 79 the cost is 6.50192029476166
At iteration 89 the cost is 4.45403056144714
At iteration 99 the cost is 2.78473098278046
```

Yeah, this is the usual behavior for the multiscale RNN.

6:10pm. I said I would stop, but let me just make sure of one thing. I will replace the LSTM activation with one composed of individual components. I am strongly suspecting that the LSTM itself is too blame, but I can only be sure of that if I get the same error curve as the one I pasted.

This won't be hard. Let me do it."