"6:15pm.
```
        inl activation_lstm memory_old {ins with input_cell output_cell forget_cell memory_cell} =
            inm f = sigmoid forget_cell
            inm i' = sigmoid input_cell
            inm o = sigmoid output_cell
            inm c' = tanh memory_cell
            inm c =
                inm x1 = hadmult (f,c)
                inm x2 = hadmult (i',c')
                add (x1, x2)
            inm h' = tanh c
            inm h = hadmult (o, h')
            succ {memory=c; out=h}
```

This should do it.

6:20pm.

```
        inm memory = tanh memory_cell

        inm a = input * memory_cell
```

Ohh, found an error. Ninja refactoring and all that. Move fast and break things.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 35.1879325866699
At iteration 19 the cost is 34.3068084716797
At iteration 29 the cost is 34.5078411102295
At iteration 39 the cost is 33.2039083480835
At iteration 49 the cost is 33.0984287261963
At iteration 59 the cost is 32.5224241256714
```

So this is what a correct LSTM would do.

Let me try the new one now that I've fixed that error."