"6:25pm.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 55.6609176635742
At iteration 19 the cost is 43.1384548187256
At iteration 29 the cost is 41.2618064880371
At iteration 39 the cost is 40.2682647705078
```

Fuck.

Though it might always be the case that the old one is wrong as well. Let me give the variables their proper names.

6:30pm.

```
        inl activation_lstm memory_old {ins with input_cell output_cell forget_cell memory_cell} =
            inm memory =
                inm input = sigmoid input_cell
                inm forget = sigmoid forget_cell
                inm memory = tanh memory_cell
                inm a = hadmult (input,memory)
                inm b = hadmult (forget,memory_old)
                add (a, b)

            inm out =
                inm memory = tanh memory
                inm output = sigmoid output_cell
                hadmult (output, memory)

            succ {memory out}
```

Let me try it like this.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 35.1879325866699
At iteration 19 the cost is 34.3068084716797
At iteration 29 the cost is 34.5078414916992
```

Yeah, it works the same as before.

From this I can conclude the the optimized LSTM activation is definitely wrong."