"6:35pm.
```
    inm out =
        inm memory_new = tanh memory
        inm output = sigmoid output_cell
        output * memory
```

Note that I am not using `memory_new` anywhere here. This might be the last error.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost is 35.2983905792236
At iteration 19 the cost is 34.9006099700928
At iteration 29 the cost is 35.3716464996338
At iteration 39 the cost is 34.6142515182495
At iteration 49 the cost is 34.7677047729492
```

I think this should be right even though the errors are not 100% the same. Yeah, it has to be.

With this I've completed the plastic LSTM. Horray!

Let me gather my thoughts."