I will finally call it a day here. No way am I going to try figuring out how to set up the gates so this thing works properly.
But thinking on it, I did have a hint that this might happen given how the MI plastic RNN performs worse than the vanilla one on this problem. I am thinking that what I should try doing is localizing the outputs.

Apart from that I need to test my hypothesis that the gating mechanisms are getting in the way here. I'll try paring down the LSTM to just two cells that multiply each other near the end.

One idea I had last night is that rather try to get the squared error which might not even be possible given that the network is doing unsupervised learning inside its inner loop (and which also has the effect of normalizing the weights,) I should relent a little and also look at the sign accuracy. This is what the Uber tests use.

That might be way to get the error to zero in a way. It might turn out the the vanilla RNN (the one with modulation) already gets this within 200 iterations.

Today was really tedious with all the type errors, but even if I do not end up using the LSTM, that activation is something I wanted to do all the way back in 2016. Back then I unfortunately did not have the capability to do that, but now I have.

Tomorrow I will as per plan spend time testing the LSTM and then after that it will be time to close the curtain on the Binary Pattern test.

I really love the plastic vanilla RNN now. I have over 10 times as much confidence over the nonmodulated one that if I use for the sake of creating adaptive players that it will work as intended. The architecture has my full confidence. It is the power that I've been seeking and have finally found. 2018 is definitely a good year for deep learning.