https://arxiv.org/abs/1809.05214 Model-Based Reinforcement Learning via Meta-Policy Optimization
I am finally ready to resume programming, but I would rather call it a day.

Today I started writing the two posts a bit past 7:30am. When I was done it was 12pm. I'd rather just go back to bed now as I am so exhausted. I have only myself to blame for this.

Also the Internet is dead so this commit will only appear later. The connection broke when Overlord was at 91%. Such unluck.

Edit: Forgot to push this an hour ago.

The paper is interesting somewhat. I am really going to have to look into what MAML is doing at some point.

Well, I roughly understand it, but I am not familiar with its details. Probably the plastic RNN that I have now vastly exceeds it in terms of capacity.

The Internet is still down.

Let me take a break here and then I will start experimentation. I think I will start off with the sign accuracy and see whether I can get anywhere with the LSTM.

The truth is, even though I only run it once, that particular performance was convincing enough to me. There really isn't anything I can do to make the LSTM perform better.

All I can do is check whether the cells are getting in the way and when I am done with that, clean my hands of it.

All that I can do at this point is maybe do an optimized implementation of the multicell RNN with the weight reversal scheme that I thought of yesterday, implement the Maze task and then add PRONG to that and then I am done with the plastic RNN arc.

That will only leave Rudder after that, but I'll delay that for a while until I find a task suitable for it.

Well, I am lying. The truth is I just do not want to deal with it right now. Now that I have something substantial I kind of want to make my first forray into the real world rather than spend yet more time on experimental methods.

It would be a different matter if Rudder was a localized learning rule like PRONG and Hebb, but as it involves building a separate pass who knows how it will turn out.

I should first spend my time gather experience with the plastic RNN. That will give me a baseline to compare Rudder against. I'll put Rudder away as a secret weapon to use in a pinch.