I am back after lying down for bit. It seems a computer reset did the trick. Now the Internet is back.
I think I figured out why the LSTM works so poorly. When would happen if I passed the inputs as identity?

Suppose there are just two cells `a` and `b`.

Suppose the output is 1 meaning `a=1,b=1`. Then `a*b=1`.
Suppose the output is 1 meaning `a=-1,b=-1`. Then `a*b=1`.

This will make for a complicated situation.

Let me run it a few times with different learning rates just so I can make sure and then I will implement sign accuracy for the regular RNN.