"3:30pm.
```
inl sign x =
    inl zero = to x 0
    if x > zero then 1
    elif x = zero then 0
    else x < zero then -1
```

This goes into the Math module.

```
The learning rate is 2 ** -12.5
At iteration 9 the cost and accuracy are 25.79309 and 0.784/1.
At iteration 19 the cost and accuracy are 23.28978 and 0.914/1.
At iteration 29 the cost and accuracy are 20.27444 and 0.958/1.
At iteration 39 the cost and accuracy are 19.74271 and 0.972/1.
At iteration 49 the cost and accuracy are 16.34487 and 0.982/1.
At iteration 59 the cost and accuracy are 13.53019 and 0.99/1.
At iteration 69 the cost and accuracy are 13.40367 and 0.986/1.
At iteration 79 the cost and accuracy are 12.68693 and 0.986/1.
At iteration 89 the cost and accuracy are 9.778257 and 0.986/1.
At iteration 99 the cost and accuracy are 7.835211 and 0.992/1.
At iteration 109 the cost and accuracy are 9.030997 and 0.98/1.
At iteration 119 the cost and accuracy are 5.432274 and 0.998/1.
At iteration 129 the cost and accuracy are 6.572613 and 0.992/1.
At iteration 139 the cost and accuracy are 5.990502 and 0.988/1.
At iteration 149 the cost and accuracy are 4.155296 and 0.998/1.
At iteration 159 the cost and accuracy are 4.902751 and 0.984/1.
At iteration 169 the cost and accuracy are 4.954143 and 0.99/1.
At iteration 179 the cost and accuracy are 4.4429 and 0.986/1.
At iteration 189 the cost and accuracy are 3.923582 and 0.986/1.
At iteration 199 the cost and accuracy are 3.351449 and 0.994/1.
At iteration 209 the cost and accuracy are 3.011658 and 1/1.
At iteration 219 the cost and accuracy are 2.567823 and 0.992/1.
At iteration 229 the cost and accuracy are 2.955546 and 0.988/1.
At iteration 239 the cost and accuracy are 2.441074 and 0.994/1.
At iteration 249 the cost and accuracy are 3.636085 and 0.984/1.
At iteration 259 the cost and accuracy are 2.087645 and 0.992/1.
At iteration 269 the cost and accuracy are 3.264758 and 0.99/1.
```

So it gets to nearly 100% in the first 60 iterations. Yeah, this is what I am talking about. It does not seem to be able to get it down perfectly. I guess I should try the multicell RNN.

But before that let me try the old version of the RNN without modulation."