3:35pm.
```
The learning rate is 2 ** -12.5
At iteration 9 the cost and accuracy are 26.44487 and 0.728/1.
At iteration 19 the cost and accuracy are 26.21615 and 0.762/1.
At iteration 29 the cost and accuracy are 26.38728 and 0.756/1.
At iteration 39 the cost and accuracy are 26.26429 and 0.764/1.
At iteration 49 the cost and accuracy are 26.29009 and 0.772/1.
At iteration 59 the cost and accuracy are 26.23435 and 0.76/1.
At iteration 69 the cost and accuracy are 26.07437 and 0.776/1.
At iteration 79 the cost and accuracy are 26.06558 and 0.786/1.
At iteration 89 the cost and accuracy are 25.92895 and 0.796/1.
At iteration 99 the cost and accuracy are 26.29171 and 0.754/1.
At iteration 109 the cost and accuracy are 26.00823 and 0.79/1.
At iteration 119 the cost and accuracy are 26.36057 and 0.76/1.
At iteration 129 the cost and accuracy are 26.11884 and 0.78/1.
At iteration 139 the cost and accuracy are 25.98312 and 0.794/1.
At iteration 149 the cost and accuracy are 25.76338 and 0.792/1.
At iteration 159 the cost and accuracy are 25.69174 and 0.812/1.
At iteration 169 the cost and accuracy are 25.74615 and 0.8/1.
At iteration 179 the cost and accuracy are 25.84141 and 0.784/1.
At iteration 189 the cost and accuracy are 25.79729 and 0.806/1.
At iteration 199 the cost and accuracy are 25.75537 and 0.812/1.
At iteration 209 the cost and accuracy are 25.53073 and 0.808/1.
```

They are not even close.

Now the multicell RNN.