The material on this branch is the work that I did on differentiable plasticity on the September of 2018.
There is a wide variety of architectures that I tried, and the `semimodulated_vanilla_oja_alt` and `multiscale_v1` work by far the best.

The work here is based on the paper 'Differentiable plasticity: training plastic neural networks with
backpropagation' by Miconi et al.

The particular architecture works about as well on the Binary Pattern test, but the two of the which I mentioned which use modulation are over an order of magnitude better than the originals.

Considering that unmodulated plastic RNNs were already competitive with architectures like SNAIL which are over 10x its size, this leads me to the conclusion that modulated plastic RNNs will in the future become by far the dominant neural net architecture. Hebbian learning is destined to become an integral part of artifical nets just like it is for biological ones.

That I am not overextrapolating from just a single test is what I need to make sure of next, though what gives me my certainty is the huge amount of preselection that Hebbian learning has in nature.

What I intend to do from here is implement the optimized implementation of the modulated plastic RNN and clean up the library of all those other unneeded architectures. After that I will move to further testing such as on the Maze task. Let this branch serve as a deposit of this moment in time.

As a brief description of modulation - input and state are mapped through a multiplicative integration operation for the feedforward and recurrent respectively and then used as the plasticity coefficients. Those plasticity coefficients are vectors of size output - one for each neuron.

This works better than merely passing input and state through a matrix and then adding them up.

I'll take the credit for the multiplicative integration part of the modulation idea.

Also unlike the unmodulated plastic RNN, having learning rates be parameters leads to instability and hence they have been replaced with hyperparameter learning rates that can be ranges.

It is possible to make these memory efficient on par with regular RNNs by using inverse weight updates on the backward pass. This is the key discovery which will assure their dominance over competing metalearning forms and for which I arrived at independently.

This will be done in the optimized implementation.