I sure have balls to make a proclamation like that in the previous commit without even testing this on anything other Binary Pattern, but then it would not be a prediction, but just statement of fact. Honestly, it converging to 99% in the first 60 iterations is good. It is unbelievably good. There is my memory that I needed right there. Unlike for standard RNNs, large patterns are actually easier for the plastic nets.
Algorithms are not just fact. What I need right now is belief. With the belief that I can make this work, the testing will happen and the steps that are necessary to make this work will come to me.

What will come next is the great cleanup. I'll get rid of all the Hebbian architectures apart from the modulated RNN.

Then I will fuse the matrix operations in that function. Then I will add PRONG. Then comes the Maze test. After that comes poker.

Let me clean up the library and then I will call it a day. Today was a 'talk a lot to myself' kind of day.